Refactor common to_level
