+++ lib/ougai.rb
@@ -2,6 +2,7 @@ require 'logger'
 require 'ougai/version'
 require 'ougai/logging'
 require 'ougai/formatters/base'
+require 'ougai/formatters/for_json'
 require 'ougai/formatters/bunyan'
 require 'ougai/formatters/readable'
 require 'ougai/formatters/pino'
