+++ lib/ougai/formatters/pino.rb
@@ -7,6 +7,8 @@ module Ougai
     # @attr [Boolean] jsonize Whether log should converts JSON
     # @attr [Boolean] with_newline Whether tailing NL should be appended
     class Pino < Base
+      include ForJson
+
       attr_accessor :jsonize, :with_newline
 
       # Intialize a formatter
@@ -41,25 +43,6 @@ module Ougai
           time: time,
           v: 1
         }.merge(data))
-      end
-
-      def to_level(severity)
-        case severity
-        when 'TRACE'
-          10
-        when 'DEBUG'
-          20
-        when 'INFO'
-          30
-        when 'WARN'
-          40
-        when 'ERROR'
-          50
-        when 'FATAL'
-          60
-        else
-          70
-        end
       end
 
       private
